[CORE] Multiple fixes: 	Removed unnecessary rounding. 	Fixed some NPE vulnerabilities. 	At server packets some variables should be static. 	Fixed some Double-checking lock. 	Minor cleanup of invalid comments. 	Fixed NSEE on AdminReload. Reported by Sternwal. 	Fixed NPE on AdminEditNpc. Reported by Rannek. 	Removed synchronizations regarding _allPlayers in L2World, FastMap is already synchronized this may lead to deathlocks. 	Attempting to fix AIOOBE in L2PcInstance. Reported by JPDist®. 	Replaced append(String) with append(char) where is possible. 	Replaced boolean literals. 	Removed unnecessary overridden methods. 	Using L suffix for longs instead of l to avoid typos. 	Some cleanup and formatting.

+++ trunk/gameserver/head-src/com/l2jfrozen/gameserver/controllers/RecipeController.java
@@ -575,7 +575,7 @@ public class RecipeController
 
 			for(L2RecipeInstance recipe : recipes)
 			{
-				int quantity = _recipeList.isConsumable() ? (int) (recipe.getQuantity() * Config.RATE_CONSUMABLE_COST) : (int) recipe.getQuantity();
+				int quantity = _recipeList.isConsumable() ? (int) (recipe.getQuantity() * Config.RATE_CONSUMABLE_COST) : recipe.getQuantity();
 
 				if(quantity > 0)
 				{
