+++ trunk/gameserver/head-src/com/l2jfrozen/gameserver/network/serverpackets/RecipeShopItemInfo.java
@@ -50,13 +50,10 @@ public class RecipeShopItemInfo extends L2GameServerPacket
 		writeD(_shopId);
 		writeD(_recipeId);
 		writeD(manufacturer != null ? (int) manufacturer.getCurrentMp() : 0);
-		writeD(manufacturer != null ? (int) manufacturer.getMaxMp() : 0);
+		writeD(manufacturer != null ? manufacturer.getMaxMp() : 0);
 		writeD(0xffffffff);
 	}
-
-	/* (non-Javadoc)
-	 * @see com.l2jfrozen.gameserver.serverpackets.ServerBasePacket#getType()
-	 */
+	
 	@Override
 	public String getType()
 	{
